offers/pyproject.toml
[tool.poetry]
name = "offers"
version = "0.1.0"
description = ""
authors = ["MarkovDV17 <markov.dv17@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
django-rest-swagger = "^2.2.0"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
